+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Nuvlabox add modal doesn't select pre-release versions
   - NuvlaBox map view now show all NuvlaBoxes, including NBs with inferred location
   - Fix refresh of cluster info on default NuvlaBox view